services/users/requirements.txt
Flask==1.0.2
flask-restplus==0.12.1
google-cloud-datastore==1.7.0

pytest==4.6.3
black==19.3b0
mock==3.0.5
flake8==3.7.7